move job entry into own file
