+++ lib/benchmark/ips.rb
@@ -2,6 +2,7 @@
 require 'benchmark/timing'
 require 'benchmark/compare'
 require 'benchmark/ips/report'
+require 'benchmark/ips/job/entry'
 require 'benchmark/ips/job'
 
 # Performance benchmarking library
