+++ lib/benchmark/ips/job.rb
@@ -10,91 +10,6 @@ module Benchmark
       # before reporting a weird runtime
       MAX_TIME_SKEW = 0.05
 
-      # Entries in Benchmark Jobs.
-      class Entry
-        # Instantiate the Benchmark::IPS::Job::Entry.
-        # @param label [#to_s] Label of Benchmarked code.
-        # @param action [String, Proc] Code to be benchmarked.
-        # @raise [ArgumentError] Raises when action is not String or not responding to +call+.
-        def initialize(label, action)
-          @label = label
-
-          if action.kind_of? String
-            compile action
-            @action = self
-            @as_action = true
-          else
-            unless action.respond_to? :call
-              raise ArgumentError, "invalid action, must respond to #call"
-            end
-
-            @action = action
-
-            if action.respond_to? :arity and action.arity > 0
-              @call_loop = true
-            else
-              @call_loop = false
-            end
-
-            @as_action = false
-          end
-        end
-
-        # The label of benchmarking action.
-        # @return [#to_s] Label of action.
-        attr_reader :label
-
-        # The benchmarking action.
-        # @return [String, Proc] Code to be called, could be String / Proc.
-        attr_reader :action
-
-        # Add padding to label's right if label's length < 20,
-        # Otherwise add a new line and 20 whitespaces.
-        # @return [String] Right justified label.
-        def label_rjust
-          label = @label.to_s
-          if label.size > 20
-            "#{label}\n#{' ' * 20}"
-          else
-            label.rjust(20)
-          end
-        end
-
-        # Call action by given times, return if +@call_loop+ is present.
-        # @param times [Integer] Times to call +@action+.
-        # @return [Integer] Number of times the +@action+ has been called.
-        def call_times(times)
-          return @action.call(times) if @call_loop
-
-          act = @action
-
-          i = 0
-          while i < times
-            act.call
-            i += 1
-          end
-        end
-
-        # Compile code into +call_times+ method.
-        # @param str [String] Code to be compiled.
-        # @return [Symbol] :call_times.
-        def compile(str)
-          m = (class << self; self; end)
-          code = <<-CODE
-            def call_times(__total);
-              __i = 0
-              while __i < __total
-                #{str};
-                __i += 1
-              end
-            end
-          CODE
-          m.class_eval code
-        end
-      end # End of Entry
-
-      # class Job
-
       # Two-element arrays, consisting of label and block pairs.
       # @return [Array<Entry>] list of entries
       attr_reader :list
